move flake8 to python requirements
